backend/requirements.txt
nexios
tortoise-orm
bcrypt
dotenv
asyncpg